+++ CHANGELOG.md
@@ -32,7 +32,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Earlier initialization strategy for Shopify integration [#6202](https://github.com/ethyca/fides/pull/6202)
 
 ### Fixed
-- Added missing "Awaiting Email Send" status to privacy request statuses [#6218](https://github.com/ethyca/fides/pull/6218)
+- Added missing "Awaiting email send" status to privacy request statuses [#6218](https://github.com/ethyca/fides/pull/6218)
 
 ## [2.63.0](https://github.com/ethyca/fides/compare/2.62.0...2.63.0)